fix: Refactor into a single file
