+++ src/functions.test.js
@@ -1,0 +1,1 @@
+const addTodo = require('./functions.js');
