--- src/modules/TaskStatus/TaskStatus.js
@@ -1,7 +1,0 @@
-export default class Status {
-  constructor(description, completed, index) {
-    this.description = description;
-    this.completed = completed;
-    this.index = index;
-  }
-}
