--- src/modules/Tasks/Tasks.js
@@ -1,9 +1,0 @@
-import addTodo from '../addTodo/addTodo.js';
-import renderTodo from '../render/renderTodo.js';
-
-const Tasks = () => {
-  addTodo();
-  renderTodo();
-};
-
-export default Tasks;
