--- src/modules/consts/consts.js
@@ -1,4 +1,0 @@
-export const todoArray = [];
-export const sendToLocalStorage = () => {
-  localStorage.setItem('todoArray', JSON.stringify(todoArray));
-};
